FLOE-196: Start of update feedback tests
Updating the feedback tests to include the requestSummary. Still need to test the interactions between the other dialogs and requestSummary. There is also an issue with running both sets of test trees together at the same time. In this commit one of them is commented out. This will be addressed in a future commit.