FLOE-196: requestSummary-feedback integration test
Tests added for testing the integration of the requestSummary dialog and the feedback tool. A couple of the tests for verifyRequests (which tests that the requstSummary model is updated correctly) have been commented out as the model listener doesn't fire. It may be due to an issue with the model relay, as the dataSourceValue model value, which contains the raw data, is updated but the transformation and update for the requests model property doesn't seem to be updated accordingly.