FLOE-230: Using model relays for match/mismatch
The isActive model values from the match and mismatch dialogs are now relayed to the feedback model via the model relay. This commit contains a bug where clicking the match button while the mismatch dialog is open will close the mismatch dialog but won't open the match dialog. It appears that this is happening because the change request for the isActive value in the click handler on the button is having no effect.